Kill macos (#5)
* build: remove macos targeting

targeting two platforms with the same codebase from the beginning is dumb, especially when
the eventuality of the macos platform for this project is in limbo. with fully featured webuis available
I'm not sure if there's a purpose in keeping it around. removed the nonfunctional macos target. iOS is
unaffected at this time.

Signed-off-by: Charles Pickering <me@charlespick.xyz>

* build: remove old separate build files

moved to a generated info.plist and moved the files out of shared into src

Signed-off-by: Charles Pickering <me@charlespick.xyz>