feat: selection persists between restarts of the app (#11)
and printers are also selectable now. the selectedPrinterIndex always keeps track of
the selected printer and is a pointer to only 1 printer in the array of configured printers.
next up: a printer struct will live in the app's kernel to hold the printer's data and coordinate
connections. when the selected printer changes, this object will be destroyed and recreated

Signed-off-by: Charles Pickering <me@charlespick.xyz>