feat: Prototype UI dashboard (#14)
The dashboard will serve as the "homescreen". Logic is in place including a stateobject to manage the connection to the printer, the active printer, and such. a TODO remains to solve the problem of updating the pointer in the array for the purposes of rendering the selected printer when an item is deleted and destroying the connection if the printer config that is deleted was the one selected.\

Signed-off-by: Charles Pickering <me@charlespick.xyz>